Consolidated auto programs
Added discriptions in swerve cause why not
