+++ src/main/java/frc/team3128/Robot.java
@@ -12,7 +12,7 @@ import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.livewindow.LiveWindow;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import frc.team3128.autonomous.AutoPrograms;
+import frc.team3128.autonomous.Trajectories;
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
@@ -22,7 +22,7 @@ public class Robot extends TimedRobot {
 
     public static RobotContainer m_robotContainer = new RobotContainer();
     private Command m_autonomousCommand;
-    public static AutoPrograms autoPrograms = new AutoPrograms();
+    public static Trajectories autoPrograms = new Trajectories();
 
     @Override
     public void robotInit(){
