+++ src/main/java/frc/team3128/autonomous/Trajectories.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Hashtable;
 
 import com.pathplanner.lib.PathConstraints;
 import com.pathplanner.lib.PathPlanner;
@@ -20,6 +21,10 @@ import edu.wpi.first.wpilibj.Filesystem;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import static frc.team3128.Constants.SwerveConstants.*;
+
+import frc.team3128.commands.CmdAlign;
+import frc.team3128.commands.CmdInPlaceTurn;
+import frc.team3128.common.narwhaldashboard.NarwhalDashboard;
 import frc.team3128.subsystems.Swerve;
 
 /**
@@ -29,8 +34,37 @@ import frc.team3128.subsystems.Swerve;
 public class Trajectories {
 
     private static HashMap<String, ArrayList<PathPlannerTrajectory>> trajectories = new HashMap<String, ArrayList<PathPlannerTrajectory>>();
+    private static HashMap<String, Command> eventMap = new HashMap<>();
 
     private static SwerveAutoBuilder builder;
+
+    public void Trajecotires(){
+        initAutoSelector();
+        initTrajectories();
+    }
+
+    private void initAutoSelector() {
+        String[] autoStrings = new String[] {};
+        NarwhalDashboard.addAutos(autoStrings);
+    }
+
+    public Command getAutonomousCommand() {
+        String selectedAutoName = NarwhalDashboard.getSelectedAutoName();
+        // String selectedAutoName = "3 Ball"; // uncomment and change this for testing without opening Narwhal Dashboard
+        selectedAutoName = ""; //"Marriage";
+
+        if (selectedAutoName == null) {
+            return null;
+        }
+
+        return builder.fullAuto(trajectories.get(selectedAutoName));
+    }
+
+    private static void initEventMap(){
+        eventMap.put("Align", new CmdAlign());
+        eventMap.put("180 Turn", new CmdInPlaceTurn(180, false));
+        eventMap.put("180 Turn Inturrputed", new CmdInPlaceTurn(180, true));
+    }
 
     public static void initTrajectories() {
         final String[] trajectoryNames = {};
@@ -51,13 +85,9 @@ public class Trajectories {
         );
     }
 
-    public static CommandBase get(String name) {
-        return builder.fullAuto(trajectories.get(name));
-    }
-
     public static PathPlannerTrajectory line(Pose2d start, Pose2d end) {
         return PathPlanner.generatePath(
-            new PathConstraints(maxSpeed, 4), 
+            new PathConstraints(maxSpeed, maxAcceleration), 
             new PathPoint(start.getTranslation(), start.getRotation()), 
             new PathPoint(end.getTranslation(), end.getRotation())
             );
